% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_support.R
\name{adjust_file_path_handling}
\alias{adjust_file_path_handling}
\title{Adjust Windows paths to R paths}
\usage{
adjust_file_path_handling(file_name_and_path)
}
\arguments{
\item{file_name_and_path}{The file path and folder to adjust.}
}
\value{
The input adjusted for R pathing.
}
\description{
Adjust Windows paths to R paths
}
\details{
Converts forward slashes to back slashes and replaces drive letter with
Data Lab location prefix. The Data Lab location prefix is specific to the
IDI at time of development.

The inclusion of this function allows researchers to copy-and-paste paths
from Windows into the pipeline control file.
}
