% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidential_support.R
\name{apply_conventional_rounding}
\alias{apply_conventional_rounding}
\title{Apply conventional rounding to an array}
\usage{
apply_conventional_rounding(input_array, base)
}
\arguments{
\item{input_array}{a numeric array to randomly round.}

\item{base}{the base to round to.}
}
\value{
the input vector with all values rounded so they are
divisible by \code{base}.
}
\description{
Apply conventional rounding to an array
}
\details{
Consistent with \code{round} in base R (which implements the IEC 60559 standard)
values of 0.5 are 'rounded to the even digit'. This is different from
simple rounding where values of 0.5 are rounded up.
}
