% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidential_support.R
\name{apply_graduated_random_rounding}
\alias{apply_graduated_random_rounding}
\title{Apply graduated random rounding (GRR) to an array}
\usage{
apply_graduated_random_rounding(
  input_array,
  seeds = NULL,
  threshold = numeric()
)
}
\arguments{
\item{input_array}{a numeric array to randomly round.}

\item{seeds}{an optional column of seed values, with length equal to
\code{input_array}. If provided, these seeds are used to determine the random
rounding - useful for consistent or reproducible rounding.}

\item{threshold}{If provided, where \code{input_array} is at least/below the
\code{threshold}, then the output will also be at least/below the threshold.
Can override randomness of rounding. Most common use case is to ensure
rounding is consistent with suppression.}
}
\value{
the input vector with all values graduated randomly rounded.
}
\description{
Apply graduated random rounding (GRR) to an array
}
\details{
Thresholds for graduated rounding are set by the Stats NZ Microdata output
guide. Take the absolute value of \code{input_array} and for each value apply
the first applicable rule:
\itemize{
\item Values less than 19 are randomly rounded to base 3
\item Values less than 20 are randomly rounded to base 2
\item Values less than 100 are randomly rounded to base 5
\item Values less than 1000 are randomly rounded to base 10
\item Values greater than 1000 are randomly rounded to base 100
}
}
