% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidential_support.R
\name{apply_small_count_suppression}
\alias{apply_small_count_suppression}
\title{Apply small count suppression to an array}
\usage{
apply_small_count_suppression(input_array, with, threshold, treat_na_as = 0)
}
\arguments{
\item{input_array}{an array to suppress.}

\item{with}{a numeric array to compare against \code{threshold} to determine
whether or not to suppress. Either length 1 or length equal to \code{input_array}.}

\item{threshold}{the minimum accepted value of \code{with} that will not result
in suppression. Either length 1 or length equal to \code{input_array}.}

\item{treat_na_as}{optional, numeric value to replace missing values.
Defaults to 0, so NA values in \code{with} will suppress \code{input_array}.
See details. Either length 1 or length equal to \code{input_array}.}
}
\value{
The \code{input_array} with values suppressed (replaced by NA) anywhere
that \code{with >= threshold} is not true.
}
\description{
Apply small count suppression to an array
}
\details{
This function takes the most conservative approach and suppresses all values
where there is not evidence that the count is at least the threshold. This
means than missing values in \code{with} will suppress \code{input_array}.

If missing values in \code{with} reflect that no suppression check is required,
then \code{treat_na_as = 9999} or similar is advised.

Missing values in \code{threshold} or \code{treat_na_as} will cause errors.
}
