% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ratio.R
\name{count_ratio}
\alias{count_ratio}
\title{Censored Regression for Count Ratios}
\usage{
count_ratio(
  otu_table,
  metadata,
  covar,
  adjust = NULL,
  refgenes = NULL,
  complement = FALSE,
  boot = TRUE,
  boot_replicate = 500,
  n_boot_gene = 100
)
}
\arguments{
\item{otu_table}{microbiome abundance matrix. All entries are integers}

\item{metadata}{sample metadata containing the covariate of interest and confounding factors for adjustment}

\item{covar}{the name of the covariate of interest}

\item{adjust}{the names of the variables for adjustment, NULL if no adjustments}

\item{refgenes}{the list of reference genes}

\item{complement}{If true, model the count ratios between each taxa beyond the reference set and the sum of taxa in the reference set. If false, model the relative abundance of taxa within the reference set.}

\item{boot}{whether to use bootstrap to estimate scaling factors for the test statistics, default TRUE}

\item{boot_replicate}{number of bootstrap replicates for each gene to estimate scaling factors}

\item{n_boot_gene}{number of genes to apply bootstrap}
}
\value{
The effect sizes, standard errors and p values for each taxon
}
\description{
Censored Regression for Count Ratios
}
