% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidential_support.R
\name{create_stable_seeds}
\alias{create_stable_seeds}
\title{Create seeds for stable random rounding}
\usage{
create_stable_seeds(source_values, stable_above = 30)
}
\arguments{
\item{source_values}{Numeric vector of values to round for which stable
random rounding is required.}

\item{stable_above}{Minimum source value for which consistent seeds are
required. Defaults to 30, because in practice smaller values come from
different sources.}
}
\value{
Numeric vector of the same length as \code{source_values} suitable for
in \code{apply_random_rounding} or \code{apply_graduated_random_rounding} as \code{seeds}.
Output values will be identical where \code{source_values} and identical and
greater than \code{stable_above}.
}
\description{
Create seeds for stable random rounding
}
