% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nova_scotia.R
\name{data_nova_scotia}
\alias{data_nova_scotia}
\title{data_nova_scotia}
\usage{
data_nova_scotia(
  output = "stan_data",
  Npop = 971395,
  Npreds = 5,
  interpolate_missing_data = FALSE,
  ...
)
}
\arguments{
\item{output}{default is \code{"stan_data"}}

\item{Npop}{default is \code{971395}.}

\item{Npreds}{default is \code{5}.}

\item{interpolate_missing_data}{default is \code{FALSE}}

\item{...}{other arguments passed to methods}
}
\value{
This is a placeholder for what it returns.
}
\description{
This is a placeholder for a description.
}
\author{
Jae Choi, \email{choi.jae.seok@gmail.com}
}
