% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_provinces_of_canada.R
\name{data_provinces_of_canada}
\alias{data_provinces_of_canada}
\title{data_provinces_of_canada}
\usage{
data_provinces_of_canada(selection = "default", fn = NULL, Npreds = 5, ...)
}
\arguments{
\item{selection}{default is \code{"default"}}

\item{fn}{default is \code{NULL}}

\item{Npreds}{default is \code{5}}

\item{...}{other arguments passed to methods}
}
\value{
This is a placeholder for what it returns.
}
\description{
This is a placeholder for a description.
}
\author{
Jae Choi, \email{choi.jae.seok@gmail.com}
}
