% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_support.R
\name{drop_sql_server_tables}
\alias{drop_sql_server_tables}
\title{Drop tables from SQL Server}
\usage{
drop_sql_server_tables(db_connection, tables, if_exists = TRUE)
}
\arguments{
\item{db_connection}{An SQL server database connection.}

\item{tables}{A character array containing a list of tables to drop.}

\item{if_exists}{T/F whether to include 'IF EXISTS' in the SQL query.
Defaults to TRUE, if FALSE the command will error if any entry in \code{tables}
does not exist in the database.}
}
\value{
An invisible numeric array of the same length as \code{tables} with 1 if
the table was dropped, and 0 otherwise.
}
\description{
Drop tables from SQL Server
}
