% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecc16s_12}
\alias{ecc16s_12}
\title{16S sequencing of saliva samples collected from 12-month-old infants}
\format{
\subsection{\code{ecc16s_12}}{

The metadata has three columns
\describe{
\item{CaseEver}{Whether the child developed dental caries after 36 months old}
\item{geo_loc_name}{Where the child lives (Pennsylvania or West Virginia)}
}
}
}
\source{
The original publication of this data is "Evaluating the ecological hypothesis: early life salivary microbiome assembly predicts dental caries in a longitudinal case-control study" (Blostein etal, 2022). The sequence data are available under Project number PRJNA752888.
}
\usage{
ecc16s_12
}
\description{
A phyloseq object with 161 samples and 280 ASVs. The samples are collected from 12-month-old infants' saliva.
84 out of 161 children developed dental caries after 36 months old. The children were from Pennsylvania and West Virginia.
}
\keyword{datasets}
