% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecc_metag}
\alias{ecc_metag}
\title{Whole genome sequencing of plaque samples collected from infants after 36 months old}
\format{
\subsection{\code{ecc_metag}}{

The metadata has three columns
\describe{
\item{CaseStatus}{Whether the kids had dental caries}
\item{Site}{Where the child lives (Pennsylvania or West Virginia)}
}
}
}
\source{
The original publication of this data is "Evaluating the ecological hypothesis: early life salivary microbiome assembly predicts dental caries in a longitudinal case-control study" (Blostein etal, 2022). The sequence data are available under Project number PRJNA752888.
}
\usage{
ecc_metag
}
\description{
A phyloseq object with 30 samples and 610 taxa. The samples are collected from plaque of children at 36, 48 or 60 months old
15 samples were from teeth with dental lesions and 15 samples were controls.
}
\keyword{datasets}
