% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecc_plaque}
\alias{ecc_plaque}
\title{Plaque samples from early childhood dental caries studies}
\format{
The metadata of \code{ecc_plaque} has two columns
\describe{
\item{CaseStatus}{Whether the kids had dental caries}
\item{Site}{The location of sample collection (Site 1 or Site 2).}
}
}
\source{
The original publication of this data is "Evaluating the ecological
hypothesis: early life salivary microbiome assembly predicts dental caries
in a longitudinal case-control study" (Blostein etal, 2022).
The sequence data are available under Project number PRJNA752888.
}
\usage{
data(ecc_plaque)
}
\description{
A phyloseq object with 30 samples and 610 taxa for whole genome sequencing of plaque samples.
The samples were collected from children at 36, 48 or 60 months old.
15 samples were from teeth with dental lesions and 15 samples were controls.
The samples of cases were collected at the onset visit.
}
\keyword{datasets}
