% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecc_saliva}
\alias{ecc_saliva}
\title{Saliva samples from early childhood dental caries studies}
\format{
The metadata of \code{ecc_saliva} has two columns
\describe{
\item{CaseStatus}{Whether the child developed dental caries after 36 months old}
\item{Site}{The location of sample collection (Site 1 or Site 2).}
}
}
\source{
The original publication of this data is "Evaluating the ecological
hypothesis: early life salivary microbiome assembly predicts dental caries
in a longitudinal case-control study" (Blostein etal, 2022).
The sequence data are available under Project number PRJNA752888.
}
\usage{
data(ecc_saliva)
}
\description{
A phyloseq object with 161 samples and 280 ASVs for 16S sequencing of saliva
samples. The samples were collected from
12-month-old infants. 84 out of 161 children developed dental caries after
36 months old. All samples have been de-identified.
}
\keyword{datasets}
