% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_support.R
\name{entity_to_min_and_max}
\alias{entity_to_min_and_max}
\title{Convert entity rows in control file to min and max rows}
\usage{
entity_to_min_and_max(control_file)
}
\arguments{
\item{control_file}{a data frame containing assembly instructions. Most
likely read into memory by \code{load_control_file}.}
}
\value{
The control file with ENTITY types removed and replaced by pairs of
MIN and MAX types.
}
\description{
Convert entity rows in control file to min and max rows
}
\details{
For the ENTITY summary method the assembly tool outputs two variables, with
\verb{*__min} and \verb{*__max} suffixes. Initial designs handled this as its own
summary method However, it is a simpler implementation to modify the control
file - replacing the ENTITY row with two rows - a MIN and a MAX. This means
the rest of the process never has to handle one row becoming two.
}
