% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_support.R
\name{entity_union_all_conversion}
\alias{entity_union_all_conversion}
\title{Take union of tbl to allow for combining of min and max entities}
\usage{
entity_union_all_conversion(summary_row, tbl)
}
\arguments{
\item{summary_row}{A single row of the control file.}

\item{tbl}{a data frame to summarise. Can be in-memory or remote accessed
with dbplyr.}
}
\value{
The tbl modified for handling \verb{*__min} and \verb{*__max} entities if
required.
}
\description{
Take union of tbl to allow for combining of min and max entities
}
\details{
The 'entity' summary type takes distinct values. It also allows for distinct
values over two columns \verb{*__min} and \verb{*__max}. The only practical way to
implement this is to take a union all of the table stacking the \verb{*__min} and
\verb{*__max} columns into a single column.
}
