% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_support.R
\name{example_control_file}
\alias{example_control_file}
\title{Notes
\itemize{
\item load_control_file needs a more general solution to column names to handle
different control file layouts
\item load_control_file needs handling for XLSX sheets
}}
\usage{
example_control_file(folder, tool)
}
\arguments{
\item{folder}{Folder to copy the example into.}

\item{tool}{Tool to fetch control file example for. Matches internal package
extdata folder. Options: assembly, summary, confidential, combined.}
}
\value{
The path of the newly created file(s)
}
\description{
Provide an example of the control file for a tool
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
example_control_file("./examples", "summary")
\dontshow{\}) # examplesIf}
}
