% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_support.R
\name{example_summary_control_file}
\alias{example_summary_control_file}
\title{Provide an example of the control file for a tool}
\usage{
example_summary_control_file(folder = ".")
}
\arguments{
\item{folder}{Folder to copy the example into. Defaults to current directory.}
}
\value{
The path of the newly created file(s) (invisibly)
}
\description{
Provide an example of the control file for a tool
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
example_summary_control_file("./examples")
\dontshow{\}) # examplesIf}
}
