% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_support.R
\name{fetch_all_sql_server_temp_tables}
\alias{fetch_all_sql_server_temp_tables}
\title{List all temporary tables in SQL Server that are associated with the
database connection.}
\usage{
fetch_all_sql_server_temp_tables(db_connection, global_temp_tables = FALSE)
}
\arguments{
\item{db_connection}{An SQL server database connection.}

\item{global_temp_tables}{T/F whether to include global temporary tables
(ones with names that start '##').}
}
\value{
Character array containing the names of all temporary tables
associated with the current connection.
}
\description{
List all temporary tables in SQL Server that are associated with the
database connection.
}
\details{
Uses the SQL query:
\verb{SELECT name FROM tempdb..sysobjects WHERE OBJECT_ID('tempdb..' + name) IS NOT NULL}
}
