% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folded_root.r
\name{folded_root}
\alias{folded_root}
\title{folded_root}
\usage{
folded_root(x, p = 1, inverse = FALSE, ...)
}
\arguments{
\item{x}{default is x}

\item{p}{default is \code{1}.}

\item{inverse}{default is \code{NULL}.}

\item{...}{other arguments passed to methods}
}
\value{
This is a placeholder for what it returns.
}
\description{
This is a placeholder for a description.
}
\author{
Jae Choi, \email{choi.jae.seok@gmail.com}
}
