% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_support.R
\name{generate_combinations_df}
\alias{generate_combinations_df}
\alias{cross_product_column_names}
\title{Creates the cross-products of its inputs as a data frame. Designed for
receiving column names as input and creating combinations to group-by in
analysis.}
\usage{
generate_combinations_df(
  ...,
  always = NULL,
  drop.dupes.within = TRUE,
  drop.dupes.across = TRUE
)

cross_product_column_names(
  ...,
  always = NULL,
  drop.dupes.within = TRUE,
  drop.dupes.across = TRUE
)
}
\arguments{
\item{...}{any number of arrays to take the cross-product between. Converted
to character. NA values in an arrays mean that the combination without this
array is also produced.}

\item{always}{an array of values that should always be included in the output
combination. Converted to character.}

\item{drop.dupes.within}{T/F, controls whether duplicated values within each
output set are discarded. For example: \verb{(a,b,c,a,b)} becomes \verb{(a,b,c)}.
Note: dplyr::group_by requires no duplicates so turning this off
may produce errors if the output is used for summarizing results.
Defaults to TRUE.}

\item{drop.dupes.across}{T/F, control whether duplicated sets of values
across the output are discarded. For example: \code{list(c(a,b,c), c(a,c,b))}
will only output \code{c(a,b,c)}.
Defaults to TRUE.}
}
\value{
A data frame produced by taking the cross-products of each input
group, appending the always inputs, and tidying by removing duplication
where applicable.
}
\description{
Creates the cross-products of its inputs as a data frame. Designed for
receiving column names as input and creating combinations to group-by in
analysis.
}
\details{
Note: because \code{...} is given as the first argument, all other arguments
need to be named. Otherwise their values will be included in the \code{...}
argument.

\code{cross_product_column_names} provides equivalent output in list format for
backwards compatibility.
}
\examples{
# basic use case
generate_combinations_df(10:11, 101:102, always = 1:3)

# use of NAs for every combination
generate_combinations_df(c("a",NA),c("b",NA),c("c",NA))

# differences when including duplicates
generate_combinations_df(1:3, 1:3)
generate_combinations_df(1:3, 1:3, drop.dupes.across = FALSE)
generate_combinations_df(1:3, 1:3, drop.dupes.within = FALSE)

# applied example - every pairwise combination of demographic columns
demo_cols = c("age", "region", "sex", "ethnicity")
generate_combinations_df(demo_cols, demo_cols)

}
