% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_count_ratio.R
\name{GLM_count_ratio}
\alias{GLM_count_ratio}
\title{Overdispersed and Bias-corrected Logistic Regression for Count Ratios}
\usage{
GLM_count_ratio(
  count_data,
  metadata,
  covar,
  adjust = NULL,
  reftaxa = NULL,
  complement = FALSE
)
}
\arguments{
\item{count_data}{microbiome abundance matrix. All entries are integers}

\item{metadata}{sample metadata containing the covariate of interest and confounding factors for adjustment}

\item{covar}{the name of the covariate of interest}

\item{adjust}{the names of the variables for adjustment, NULL if no adjustments}

\item{reftaxa}{the list of reference taxa}

\item{complement}{If true, model the count ratios between each taxa beyond the reference set and the sum of taxa in the reference set. If false, model the relative abundance of taxa within the reference set.}
}
\value{
The effect sizes, standard errors and p values for each taxon
}
\description{
Overdispersed and Bias-corrected Logistic Regression for Count Ratios
}
