% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_support.R
\name{handle_summary_case}
\alias{handle_summary_case}
\title{Handle summary cases for assembly}
\usage{
handle_summary_case(control_file_row, sqlite = FALSE)
}
\arguments{
\item{control_file_row}{A row on an assembly control_file.}

\item{sqlite}{T/F is the connection SQLite (different handling is required
as SQLite does not have a DATE data type).}
}
\value{
Text containing the SQL calculation. When method is ENTITY, this
text will have length = 2.
}
\description{
A range of summary methods can be chosen for the output:
MIN, MAX, EXISTS, COUNT, MEAN, DISTINCT, ENTITY, SUM, SUM_WITHIN, DURATION.
}
\details{
Each of these needs to be translated into an SQL calculation for the form:
\verb{method(input_column) AS output_column}.
These are used to build the SQL code required for assembly.
}
