% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{increment_file_name}
\alias{increment_file_name}
\title{Increment number in file path to create a unique file name}
\usage{
increment_file_name(path_and_file_name)
}
\arguments{
\item{path_and_file_name}{The path to the file that might need incrementing.}
}
\value{
The file path incremented if necessary to ensure file name is unique.
}
\description{
Increment number in file path to create a unique file name
}
\details{
This is a simple mimic of an existing Windows function that will rename
\code{file.csv} to \verb{file (1).csv} or \verb{file (1).csv} to \verb{file (2).csv} if the
existing file name already exists.
}
