% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{is_delimited}
\alias{is_delimited}
\title{Check string for delimiter}
\usage{
is_delimited(string, delimiter)
}
\arguments{
\item{string}{the string to check for delimiter}

\item{delimiter}{a 1 or 2 character string containing the delimiter}
}
\value{
T/F if the string is delimited
}
\description{
The entries in the input control tables should be delimited as either
[] for sql columns or "" for strings
This lets us run a check for the right delimiter, e.g.
is_delimited(string, "[]")
is_delimited(string, "\"")
}
