% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_support.R
\name{merge_progress_into_control_file}
\alias{merge_progress_into_control_file}
\title{Merge progress information into control file}
\usage{
merge_progress_into_control_file(control_file, progress_df)
}
\arguments{
\item{control_file}{A loaded control file to update.}

\item{progress_df}{A data frame with the columns start_time, end_time, and
status. And all other columns in common with the control file.}
}
\value{
The control file with progress information merged in.
}
\description{
Merge progress information into control file
}
\details{
Merging only occurs when rows are enabled. Rows that have been disabled are
not updated.
}
