% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{no_obvious_escaping_injection}
\alias{no_obvious_escaping_injection}
\title{Check input string for obvious markers of code injection.}
\usage{
no_obvious_escaping_injection(string)
}
\arguments{
\item{string}{a string to be checked for special characters.}
}
\value{
TRUE if no special characters in string and any quotes are matched,
FALSE if special characters or unmatched quotes found.
}
\description{
Identifies special characters \verb{;\{\}}, unmatched quotes, and unmatched
brackets. Helps prevent escaping code injection, but not a complete solution.
}
