% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_fit.R
\name{plot_model_fit}
\alias{plot_model_fit}
\title{plot_model_fit}
\usage{
plot_model_fit(
  selection = "default",
  stan_results = NULL,
  sim = NULL,
  nsimlines = 2000,
  outdir = "",
  nx = NULL,
  to.screen = FALSE
)
}
\arguments{
\item{selection}{default is sim}

\item{stan_results}{default is \code{NULL}}

\item{sim}{default is \code{NULL}}

\item{nsimlines}{default is \code{2000}}

\item{outdir}{default is \code{""}}

\item{nx}{default is \code{NULL}}

\item{to.screen}{default is \code{FALSE}}
}
\value{
This is a placeholder for what it returns.
}
\description{
This is a placeholder for a description.
}
\author{
Jae Choi, \email{choi.jae.seok@gmail.com}
}
