% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLDA.R
\name{polda}
\alias{polda}
\title{Pooling overdispersed logistic regressions for microbiome differential abundance analysis}
\usage{
polda(
  otu_table,
  metadata,
  covar,
  adjust = NULL,
  prevalence_cutoff = 0.1,
  depth_cutoff = 1000,
  features_are_rows = TRUE,
  ratio_model = c("loglogistic", "weibull"),
  alpha = 0.05
)
}
\arguments{
\item{otu_table}{microbiome abundance table generated from 16S rRNA sequencing or shotgun metagenomic sequencing. All entries are integers}

\item{metadata}{sample metadata dataframe}

\item{covar}{the name of the covariate of interest}

\item{adjust}{the names of confounders to adjust for, NULL if no adjustment}

\item{prevalence_cutoff}{features whose prevalence are smaller than the cutoff will be excluded from analysis}

\item{depth_cutoff}{a sample would be discarded if its sequencing depth is smaller than the threshold}

\item{features_are_rows}{whether the microbiome features are on the column or the row of the abundance table}

\item{alpha}{the cutoff of the adjusted p values}
}
\value{
the reference feature set, the identified DA features and the p values for all the features
}
\description{
Pooling overdispersed logistic regressions for microbiome differential abundance analysis
}
