% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.adapt}
\alias{predict.adapt}
\title{Make predictions from an adaptive lasso model.}
\usage{
\method{predict}{adapt}(obj, newx)
}
\arguments{
\item{obj}{an adaptive lasso model}

\item{newx}{the covariate values to use for prediction}
}
\description{
Make predictions from an adaptive lasso model.
}

