% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess input phyloseq data to get count table and metadata design matrix}
\usage{
preprocess(input_data, cond.var, base.cond, adj.var, prev.filter, depth.filter)
}
\arguments{
\item{input_data}{a phyloseq object}

\item{cond.var}{the variable representing the conditions to compare, a character string}

\item{base.cond}{the condition chosen as baseline. This is only used when the condition is categorical.}

\item{adj.var}{the names of the variables to be adjusted, a vector of character strings}

\item{prev.filter}{taxa whose prevalences are smaller than the cutoff will be excluded from analysis}

\item{depth.filter}{a sample would be discarded if its library size is smaller than the threshold}
}
\value{
a list with count table, metadata design matrix and DAA name
}
\description{
Preprocess input phyloseq data to get count table and metadata design matrix
}
