% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTDA.R
\name{ptda}
\alias{ptda}
\title{Pooling Tobit Models for microbiome differential abundance analysis}
\usage{
ptda(
  otu_table,
  metadata,
  covar,
  adjust = NULL,
  prevalence_cutoff = 0.1,
  depth_cutoff = 1000,
  features_are_rows = TRUE,
  ratio_model = c("lognormal", "loglogistic"),
  zero_censor = 1,
  firth = T,
  pen = 0.5,
  alpha = 0.05
)
}
\arguments{
\item{otu_table}{microbiome abundance table generated from 16S rRNA sequencing or shotgun metagenomic sequencing. All entries are integers}

\item{metadata}{sample metadata dataframe}

\item{covar}{the name of the covariate of interest}

\item{adjust}{the names of confounders to adjust for, NULL if no adjustment}

\item{prevalence_cutoff}{features whose prevalence are smaller than the cutoff will be excluded from analysis}

\item{depth_cutoff}{a sample would be discarded if its sequencing depth is smaller than the threshold}

\item{features_are_rows}{whether the microbiome features are on the column or the row of the abundance table}

\item{ratio_model}{the censored regression model for the count ratios. lognormal, loglogistic or weibull}

\item{zero_censor}{the nonzero value to replace zeros when calculating censored count ratios, default 1}

\item{firth}{whether to add firth penalty to the likelihood}

\item{pen}{penalty coefficient for log likelihood, default 0.5}

\item{alpha}{the cutoff of the adjusted p values}
}
\value{
the reference feature set, the identified DA features and the p values for all the features
}
\description{
Pooling Tobit Models for microbiome differential abundance analysis
}
