% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_support.R
\name{read_and_prepare_sql_code}
\alias{read_and_prepare_sql_code}
\title{Read in an SQL file and prepare for execution via ODBC}
\usage{
read_and_prepare_sql_code(file_name_and_path)
}
\arguments{
\item{file_name_and_path}{the location of the SQL file to read.}
}
\value{
A list with three components of equal length, where the length of
each component matches the number of batches in the code file. The components
are:
\itemize{
\item \code{code} containing the code for the batch.
\item \code{start_lines} containing the line number of the first line of the batch.
\item \code{end_lines} containing the line number of the last line of the batch.
}
}
\description{
Read in an SQL file and prepare for execution via ODBC
}
\details{
Preparation for ODBC includes: (1) removing comments, and (2) splitting code
into batches by breaking on \verb{;} and by \code{GO}.
}
