% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_support.R
\name{remove_empty_entity_columns}
\alias{remove_empty_entity_columns}
\title{Remove empty entity columns}
\usage{
remove_empty_entity_columns(summary_control_file, tbl)
}
\arguments{
\item{summary_control_file}{a loaded data frame containing summary
instructions.}

\item{tbl}{the data frame to be summarised. Can be in-memory or remote
accessed with dbplyr.}
}
\value{
The control file with entity cells emptied if the entity column
contains no entries.

Because we handled two entity columns (\verb{*_min} and \verb{*_max}) summarising
entities tends to be slow in SQL due to how this is implemented. Changing the
implementation is not straightforward given the need to handle pairs of
entity columns. But we can remove entity columns from the control file where
the column in \code{tbl} contains no values.
}
\description{
Remove empty entity columns
}
