% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_support.R
\name{remove_sql_comments}
\alias{remove_sql_comments}
\title{Locate and remove in-line and multi-line SQL comments}
\usage{
remove_sql_comments(code)
}
\arguments{
\item{code}{A character string contaiing the SQL code.}
}
\value{
Commentless code, a character string with all the comments removed
but no changes to newline characters so the lines are identical.
}
\description{
Locate and remove in-line and multi-line SQL comments
}
\details{
SQL have two types of comments. in-line comments start with \verb{--} and end at
the end of the line. Multi-line comments start with \verb{/*} and end with \verb{*/}.

The complexity of removing comments arises when these types interact. For
example, how should the line \verb{-- */ text } be understood - is the text within
a comment or not?

Testing in SQL Server shows that where comments interact, the comment that
starts first lasts until it ends and any comment-markers within it are
ignored. To handle this we need to scan through the code string rather than
using simple find and replace.
}
