% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_support.R
\name{save_control_file_w_progress}
\alias{save_control_file_w_progress}
\title{Save progress update to control file}
\usage{
save_control_file_w_progress(
  path_and_file_name,
  sheet = NULL,
  progress_df,
  overwrite = TRUE
)
}
\arguments{
\item{path_and_file_name}{location of the file to read into R}

\item{sheet}{Sheet to read if Excel file as per \code{openxlsx2::read_xlsx}:
Either a string (name of a sheet), or an integer (the position of the
sheet). Defaults to the first sheet otherwise.}

\item{progress_df}{A data frame with the columns start_time, end_time, and
status. And all other columns in common with the control file.}

\item{overwrite}{T/F should the existing file be overwritten. If FALSE, will
increment the file name (for example: \code{file.csv} becomes \verb{file (1).csv}).
If TRUE will attempt to overwrite and if overwrite fails, will increment the
file name.}
}
\value{
The path of the written file.
}
\description{
Save progress update to control file
}
