% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{simulate}
\usage{
simulate(
  stan_results,
  istart = NULL,
  nsims = 1,
  nprojections = 10,
  nthreads = 1,
  model = "stochastic.simulation.sir"
)
}
\arguments{
\item{stan_results}{default is}

\item{istart}{default is \code{NULL}.}

\item{nsims}{default is \code{1}.}

\item{nprojections}{default is \code{10}}

\item{nthreads}{default is \code{1}}

\item{model}{default is \code{"stochastic.simulation.sir"}}
}
\value{
This is a placeholder for what it returns.
}
\description{
This is a placeholder for a description.
}
\author{
Jae Choi, \email{choi.jae.seok@gmail.com}
}
