% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_add_noise.R
\name{simulate_add_noise}
\alias{simulate_add_noise}
\title{simulate_add_noise}
\usage{
simulate_add_noise(sim, nobs = 30)
}
\arguments{
\item{sim}{default is sim}

\item{nobs}{default is \code{30}}
}
\value{
This is a placeholder for what it returns.
}
\description{
This is a placeholder for a description.
}
\author{
Jae Choi, \email{choi.jae.seok@gmail.com}
}
