% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{simulate_data}
\usage{
simulate_data(
  selection = "sir",
  Npop = 1,
  inits = c(0.99, 0.01, 0),
  times = 0:99,
  params = list(beta = 0.6, gamma = 0.1),
  plotdata = TRUE,
  sample_init = NULL,
  nsample = NULL
)
}
\arguments{
\item{selection}{default is \code{"sir"}}

\item{Npop}{default is \code{1}}

\item{inits}{default is \code{c(0.99, 0.01, 0)}}

\item{times}{default is \code{0:99}}

\item{params}{default is \code{list(beta=0.6, gamma=0.1)}}

\item{plotdata}{default is \code{TRUE}}

\item{sample_init}{default is \code{NULL}}

\item{nsample}{default is \code{NULL}}
}
\value{
This is a placeholder for what it returns.
}
\description{
This is a placeholder for a description.
}
\author{
Jae Choi, \email{choi.jae.seok@gmail.com}
}
