% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_stan_model_code.R
\name{sir_stan_model_code}
\alias{sir_stan_model_code}
\title{sir_stan_model_code}
\usage{
sir_stan_model_code(selection = "default")
}
\arguments{
\item{selection}{default is \code{"default"}}
}
\value{
This is a placeholder for what it returns.
}
\description{
This is a placeholder for a description.
}
\author{
Jae Choi, \email{choi.jae.seok@gmail.com}
}
