% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sql2id}
\alias{sql2id}
\title{Convert character string of SQL name to Id}
\usage{
sql2id(sql_string)
}
\arguments{
\item{sql_string}{A string containing the name of an SQL object, most likely
delimited with square brackets.}
}
\value{
A \code{DBI::Id} object converting from splitting \code{sql_string} by \code{.}
and removing \verb{[]} delimiters.
}
\description{
Convert character string of SQL name to Id
}
