% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_support.R
\name{sql_file_exists_and_contains}
\alias{sql_file_exists_and_contains}
\title{Check whether an SQL file exists and contains the specified text.}
\usage{
sql_file_exists_and_contains(file, text)
}
\arguments{
\item{file}{Path to the file to check for existence and contents.}

\item{text}{Text that needs to found within the file. Special characters
are taken as literal rather than as regular expression. Characters
\code{[} and \verb{]} are made optional.}
}
\value{
T/F Whether the file exists and contains the specified text.
}
\description{
Used to estimate whether an SQL code file will create a table named \code{text}
or containing a column \code{text}.
}
