% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_adapt.r
\name{summary_adapt}
\alias{summary_adapt}
\title{summary_adapt}
\usage{
summary_adapt(
  selection = "summary.load",
  aus,
  fn = NULL,
  workdir = getwd(),
  modelname = "default",
  brks = 30,
  to.screen = TRUE
)
}
\arguments{
\item{selection}{default is \code{"summary.load"}}

\item{aus}{default is aus}

\item{fn}{default is \code{NULL}.}

\item{workdir}{default is \code{getwd()}}

\item{modelname}{default is \code{"default"}}

\item{brks}{default is \code{30}}

\item{to.screen}{default is \code{TRUE}}
}
\value{
This is a placeholder for what it returns.
}
\description{
This is a placeholder for a description.
}
\author{
Jae Choi, \email{choi.jae.seok@gmail.com}
}
