% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_support.R
\name{tolower_colnames}
\alias{tolower_colnames}
\title{Convert column names to lower case}
\usage{
tolower_colnames(tbl)
}
\arguments{
\item{tbl}{a data frame to rename. Can be in-memory or remote accessed
with dbplyr.}
}
\description{
R is case sensitive, but most SQL databases are case insensitive, and user
input in control files is not guaranteed to be case sensitive. One simple
solution is to convert all column names to lower case. This function does so
in a way that is dbplyr compatible for remote tables.
}
