% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_support.R
\name{try_run_R_file}
\alias{try_run_R_file}
\title{Execute R script in separate environment}
\usage{
try_run_R_file(file, ignore_warnings = FALSE)
}
\arguments{
\item{file}{The file (and path) of the R script to execute. Errors if file
does not exist or is not a .R file.}

\item{ignore_warnings}{T/F whether execution should continue even if warnings
occur. If \code{FALSE} (the default) then will stop on the first warning and
return the warning message. If \code{TRUE} then will suppress all warnings.}
}
\value{
A list containing three components:
\itemize{
\item status - Status message, of success or stopped with error/warning and the
error/warning message.
\item start_time - the system time when the file started running.
\item end_time - the system time when the file ceased running.
}
}
\description{
Execute R script in separate environment
}
\details{
The R script is executed in a new base environment. This means that variables
in the calling environment (e.g. within the pipeline tool) can not be used
or changed by the \code{file}.

If you want to run multiple R scripts that interact (for example a setup
script followed by a processing script), you can not run them all via this
function because each will be executed in an isolated environment.

The best alternative is to create an overview script that runs all the
interacting scripts, and use this function to run just this overview script.
}
