% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_run.R
\name{try_run_SQL_query}
\alias{try_run_SQL_query}
\title{Execute SQL query with error handling}
\usage{
try_run_SQL_query(query, db_connection, ignore_warnings = FALSE)
}
\arguments{
\item{query}{The query to execute}

\item{db_connection}{A connection to the database where assembly is to occur.}

\item{ignore_warnings}{T/F whether execution should continue even if warnings
occur. If \code{FALSE} (the default) then will stop on the first warning and
return the warning message. If \code{TRUE} then will suppress all warnings.}
}
\value{
A list containing three components:
\itemize{
\item status - Status message, of success or stopped with error/warning and the
error/warning message.
\item start_time - the system time when the query started running.
\item end_time - the system time when the query ceased running.
}
}
\description{
Execute SQL query with error handling
}
\details{
The SQL query is executed in the database environment on the provided
connection.

It is assumed that \code{query} is a single SQL query. This is not validated and
multi-query submissions may work, but are not supported. The internal
function \code{try_run_SQL_file} is designed to handle entire SQL scripts and may
be suitable for multi-query applications.
}
