% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_validate.R
\name{validate_assembly_control_file}
\alias{validate_assembly_control_file}
\title{Notes}
\usage{
validate_assembly_control_file(
  control_file,
  connection_string,
  master_table,
  sql_folder = "."
)
}
\arguments{
\item{control_file}{a data frame containing assembly instructions. Most
likely read into memory by \code{load_control_file}.}

\item{connection_string}{A connection string for the database containing
the data, where assembly should take place.}

\item{master_table}{The name of the table onto which columns should be
assembled. It is recommended using the full table name: database.schema.table}

\item{sql_folder}{The folder location containing SQL scripts. Will only be
used if tables listed in the control file can not be found in the database
and might have to be generated from SQL scripts.}
}
\value{
T/F whether or not all validation checks are passed. Generating
warnings for all failed checks.
}
\description{
Confirm that control file instructions can be executed.
}
\details{
The following checks are run and generate a failure if not passed:
\itemize{
\item Required column names are present.
\item Inputs match expected formats / delimiters.
\item Master table exists.
\item Master table columns exist.
\item Measure tables exist in the database or are found in the measure_file.
\item Measure columns exist in the database or are found in the measure_file.
\item Output methods are accepted types.
}

The following checks are run and only generate a warning if not passed:
\itemize{
\item Unrecognised column names are present.
}
}
