% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_support.R
\name{warn_if_excel_lockfile}
\alias{warn_if_excel_lockfile}
\title{Warn if Excel lock file detected}
\usage{
warn_if_excel_lockfile(path_and_file_name, delay_seconds = 10)
}
\arguments{
\item{path_and_file_name}{location and name of the control file.}

\item{delay_seconds}{number of seconds to wait after warning.}
}
\value{
Logical whether the control file exists (invisibly).
}
\description{
To reduce the likelihood of control files being edited but not saved, this
function warns if an Excel lock file is detected. It then waits 10 seconds
before continuing. This allows users to save and close the control file if
it was left open by mistake, or to ignore (and wait 10 seconds) if the file
is open deliberately.
}
