% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADAPT.R
\name{adapt}
\alias{adapt}
\title{Pooling Tobit Models for microbiome differential abundance analysis}
\usage{
adapt(
  input_data,
  metadata,
  cond.var,
  ref.cond = NULL,
  adj.var = NULL,
  prev.filter = 0.05,
  depth.filter = 1000,
  alpha = 0.05
)
}
\arguments{
\item{input_data}{a phyloseq object}

\item{cond.var}{the variable representing the conditions to compare, a character string}

\item{ref.cond}{the condition chosen as baseline. This is only used when the condition is categorical.}

\item{adj.var}{the names of the variables to be adjusted, a vector of character strings}

\item{prev.filter}{taxa whose prevalences are smaller than the cutoff will be excluded from analysis, default 0.05}

\item{depth.filter}{a sample would be discarded if its library size is smaller than the threshold}

\item{alpha}{the cutoff of the adjusted p values}
}
\value{
the reference taxa set, the identified differentially abundant taxa and the p values for all the taxa
}
\description{
Pooling Tobit Models for microbiome differential abundance analysis
}
