% Generated by roxygen2 (4.0.1): do not edit by hand
\name{adapt}
\alias{adapt}
\title{Fit an adaptive lasso model.}
\usage{
adapt(formula, data, adapt = TRUE, selection.method = "AICc",
  selectonly = FALSE)
}
\arguments{
\item{formula}{A formula listing the response and covariates}

\item{data}{A data.frame containing the data to use in building the model.}

\item{adapt}{Whether or not to use adaptive weights in the lasso}

\item{selection.method}{The selection criterion to minimize in model fitting}

\item{selectonly}{Use the adaptive lasso only for selection, doing coefficient estimation via OLS?}
}
\description{
This function produces a linear regression model via the adaptive lasso. Given a formula, a data.frame, and a selection.method, \code{adapt} will fit a linear model for the predictor in the formula, using the data, that minimizes the selection.method.
}

