% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_support.R
\name{alter_table_add_column}
\alias{alter_table_add_column}
\title{Create alter table statement to add columns}
\usage{
alter_table_add_column(table_name, columns, types)
}
\arguments{
\item{table_name}{The name of the table to which columns should be added.
It is recommended to use the full table name: database.schema.table}

\item{columns}{An array containing the name of the column or columns to add.}

\item{types}{An array of the SQL data types that correspond to columns.
Will error if invalid SQL data types are provided.}
}
\value{
A character string with an SQL Server query for adding columns.
}
\description{
Create alter table statement to add columns
}
