% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_support.R
\name{alter_table_drop_column}
\alias{alter_table_drop_column}
\title{Notes}
\usage{
alter_table_drop_column(table_name, columns, if_exists = TRUE)
}
\arguments{
\item{table_name}{The name of the table from which columns should be dropped.
It is recommended to use the full table name: database.schema.table}

\item{columns}{An array containing the name of the column or columns to drop.}

\item{if_exists}{T/F whether the query should account for existence of
columns. If TRUE, adds text "IF EXISTS" to result. If FALSE, executing the
query will error if any of the listed columns are not present.}
}
\value{
A character string with an SQL Server query for dropping columns.
}
\description{
Create alter table statement to drop columns
}
