% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidential_support.R
\name{apply_random_rounding}
\alias{apply_random_rounding}
\title{Apply random rounding to an array}
\usage{
apply_random_rounding(
  input_array,
  base = 3,
  seeds = NULL,
  threshold = numeric()
)
}
\arguments{
\item{input_array}{a numeric array to randomly round.}

\item{base}{the base to round to. defaults to 3. (RR3)}

\item{seeds}{an optional column of seed values, with length equal to
\code{input_array}. If provided, these seeds are used to determine the random
rounding - useful for consistent or reproducible rounding.}

\item{threshold}{If provided, where \code{input_array} is at least/below the
\code{threshold}, then the output will also be at least/below the threshold.
Can override randomness of rounding. Most common use case is to ensure
rounding is consistent with suppression.}
}
\value{
the input vector with all values randomly rounded so they are
divisible by \code{base}.
}
\description{
Apply random rounding to an array
}
\details{
Randomly rounds a numeric vector to the given base.
For example, if base is 3 (default) then:
\itemize{
\item 3 & 6 will be left as is because they are already base 3;
\item 4 & 7 will be rounded down with probability 2/3 and up with probability 1/3;
\item 5 & 8 will be rounded down with probability 1/3 and up with probability 2/3.
}
}
