% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ratio.R
\name{count_ratio}
\alias{count_ratio}
\title{Censored Regression for Count Ratios}
\usage{
count_ratio(
  count_data,
  metadata,
  covar,
  adjust = NULL,
  reftaxa = NULL,
  complement = FALSE,
  ratio_model = c("lognormal", "loglogistic"),
  zero_censor = 1,
  firth = T,
  pen = 0.5
)
}
\arguments{
\item{count_data}{microbiome abundance matrix. All entries are integers}

\item{metadata}{sample metadata containing the covariate of interest and confounding factors for adjustment}

\item{covar}{the name of the covariate of interest}

\item{adjust}{the names of the variables for adjustment, NULL if no adjustments}

\item{reftaxa}{the list of reference taxa}

\item{complement}{If true, model the count ratios between each taxa beyond the reference set and the sum of taxa in the reference set. If false, model the relative abundance of taxa within the reference set.}

\item{ratio_model}{Use lognormal, loglogistic model or the Weibull model to model count ratios with excessive zeros}

\item{zero_censor}{the nonzero value to replace zeros when calculating censored count ratios, default 1}

\item{firth}{whether to add firth penalty to the likelihood}

\item{pen}{penalty coefficient, default 0.5}
}
\value{
The effect sizes, standard errors and p values for each taxon
}
\description{
Censored Regression for Count Ratios
}
