% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ratio.R
\name{count_ratio}
\alias{count_ratio}
\title{Censored Regression for Count Ratios}
\usage{
count_ratio(
  count_table,
  design_matrix,
  reftaxa = NULL,
  censor = 1,
  test_all = FALSE
)
}
\arguments{
\item{count_table}{microbiome abundance matrix. All entries are integers}

\item{design_matrix}{the design matrix of covariates}

\item{reftaxa}{the list of reference taxa}

\item{censor}{the value to censor at for zero counts}

\item{test_all}{If true, model the count ratios between all taxa against the reference set; otherwise, only model the relative abundance within the reference set}
}
\value{
The effect sizes, standard errors and p values for each taxon
}
\description{
Censored Regression for Count Ratios
}
