% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dfr}
\alias{dfr}
\title{dfr}
\usage{
dfr(data, covar, tpair, indveff = NULL)
}
\arguments{
\item{data}{dataframe}

\item{covar}{name of covariates}

\item{tpair}{names of taxa pairs}

\item{indveff}{name of random effect level, optional}
}
\value{
logistic regression result
}
\description{
Differential Ratio Analysis
}
