% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dfr}
\alias{dfr}
\title{dfr}
\usage{
dfr(phyloseq_data, tpair, covar, adjust = c())
}
\arguments{
\item{phyloseq_data}{phyloseq data}

\item{tpair}{names of taxa pairs}

\item{covar}{name of covariates of interest}

\item{adjust}{names of covariates of no interest}
}
\value{
logistic regression result
}
\description{
Differential Ratio Analysis
}
