% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_support.R
\name{example_control_file}
\alias{example_control_file}
\title{Notes
\itemize{
\item Need provide_worked_examples function, or merge with example_control_file
for general provide_example function
}}
\usage{
example_control_file(folder, tool)
}
\arguments{
\item{folder}{Folder to copy the example into.}

\item{tool}{Tool to fetch control file example for. Matches internal package
extdata folder. Options: assembly, summary, confidential, combined.}
}
\value{
The path of the newly created file(s)
}
\description{
Provide an example of the control file for a tool. General version for reuse
across tools.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
example_control_file("./examples", "summary")
\dontshow{\}) # examplesIf}
}
