% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_support.R
\name{expand_compact_summary_groups}
\alias{expand_compact_summary_groups}
\title{Expand compact summary notation}
\usage{
expand_compact_summary_groups(
  compact_control_file,
  column_names = character(0),
  drop.dupes.within = TRUE,
  drop.dupes.across = TRUE
)
}
\arguments{
\item{compact_control_file}{a data frame containing a control file that may
have compact notation (see details) for expansion.}

\item{column_names}{a character array containing column names of the table.
Used to expand \verb{*_suffix} and \verb{prefix_*} notation. Not required if such
notation is not used.}

\item{drop.dupes.within}{T/F, controls whether duplicated values within each
output set are discarded. For example: \verb{(a,b,c,a,b)} becomes \verb{(a,b,c)}.
Note: dplyr::group_by requires no duplicates so turning this off
may produce errors if the output is used for summarizing results.
Defaults to TRUE.}

\item{drop.dupes.across}{T/F, control whether duplicated sets of values
across the output are discarded. For example: \code{list(c(a,b,c), c(a,c,b))}
will only output \code{c(a,b,c)}.
Defaults to TRUE.}
}
\value{
a data frame containing a control file with any compact notation
expanded. So what was one row in the compact input is multiple output rows.
}
\description{
By requiring users to list all combinations in the summary control file, the
tool maximizes transparency: a user looking at the control file can see
exactly what will be run. The downside of this is that the control file can
become very long.
}
\details{
This function provides an alternative way to generate a summary control file.
It accepts compact notation and expands it to multiple rows.#'

Three types of compact notation are accepted:
\enumerate{
\item The or: \code{|}. Text in a group column separated by \code{|} is split to create
create new rows, one for each option. So \code{a|b} will lead to two output rows,
one grouped by \code{a} and the other by \code{b}.
\item The no-group: \verb{|+}. Text in a group column with \verb{|+} on the end will also
create the ungrouped option. So \verb{a|+} will lead to two output rows, one
grouped by \code{a} and the other without a value.
\item Prefix and suffix pattern matching: \verb{*_suffix} and \verb{prefix_*}. Text in a
group column with either of these patterns will search for all options in
\code{column_names} that follow the same pattern. This is equivalent to listing
all the columns separated by \code{|} when the columns all share the same prefix
or suffix. Note that prefixes and suffixes must include an underscore
next to the \code{*}.
}
}
