% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_support.R
\name{generate_summary_commands}
\alias{generate_summary_commands}
\title{Convert control file input to text string that can be used within
dplyr::mutate via rlang::parse_exprs.}
\usage{
generate_summary_commands(summary_row)
}
\arguments{
\item{df}{a data frame containing 1 row.}
}
\value{
a named character array suitable for parse_exprs
}
\description{
Processes columns of types: "distinct", "count", "sum", "entity", "stddev"
}
