% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_support.R
\name{generate_summary_commands}
\alias{generate_summary_commands}
\title{Convert control file input to summary commands.}
\usage{
generate_summary_commands(summary_row, is_sql = FALSE)
}
\arguments{
\item{summary_row}{a data frame containing 1 row.}

\item{is_sql}{T/F whether summary occurs in SQL context or not. Counting
distinct values in SQL always ignores missing values. Hence for
consistency \code{n_distinct} uses argument \code{na.rm = TRUE} if not SQL and no
argument if SQL (as this argument is not handled during SQL translation).}
}
\value{
a named character array suitable for parse_exprs
}
\description{
These summary commands are text string that can be used within
rlang::parse_exprs within dplyr::mutate.
}
\details{
Processes columns of types: "distinct", "count", "sum", "entity", "stddev",
and "seed".
}
