% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Negative of \%in\%}
\usage{
x \%not_in\% y
}
\arguments{
\item{x}{an array of elements}

\item{y}{an array of elements}
}
\value{
T/F for each element in \code{x} whether the element does NOT appear in
\code{y}. Conversion may occur implicitly between data types.
}
\description{
Negative of \%in\%
}
