% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_support.R
\name{handle_delimiters_and_prefixes}
\alias{handle_delimiters_and_prefixes}
\title{Handle delimiters and SQL prefixes when preparing core assembly query}
\usage{
handle_delimiters_and_prefixes(
  df,
  mt_prefix,
  mt_cols,
  measure_prefix,
  measure_cols
)
}
\arguments{
\item{df}{A data frame extracted from an assembly control file.}

\item{mt_prefix}{The prefix to add to columns found in \code{mt_cols}.}

\item{mt_cols}{List of columns that should receive \code{mt_prefix}.}

\item{measure_prefix}{The prefix to add to columns found in \code{measure_cols}.}

\item{measure_cols}{List of columns that should receive \code{measure_prefix}.}
}
\value{
The input data frame prepared for using in making SQL queries.
}
\description{
Handle delimiters and SQL prefixes when preparing core assembly query
}
\details{
Modifies columns with the names: period_start, period_end, measure_start,
measure_end, and measure_value according to the following rules:
\itemize{
\item Inputs delimited with \verb{"} becomes delimited by \verb{'}. This changes constants
in the control file to text strings for SQL.
\item Inputs delimited with \code{{}} have their delimiters removed. This changes
dynamic input in the control file to pure SQL.
\item Columns in \code{mt_cols} are given the prefix \code{mt_prefix}. This makes the
source from which columns are fetched in the core query clear.
\item Columns in \code{measure_cols} are given the prefix \code{measure_prefix}. This makes
the source from which columns are fetched in the core query clear.
}

The first pair of tasks are necessary because we are moving between
environments with different delimiters. The second pair of tasks are
necessary to avoid confusion when input tables have the same column name.
Regex approach required to handle cases where control_file contents are
constant or dynamic.
}
