% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_support.R
\name{is_valid_data_type}
\alias{is_valid_data_type}
\title{Check whether a data type is a valid SQL Server data type}
\usage{
is_valid_data_type(type)
}
\arguments{
\item{type}{The data type to test.}
}
\value{
T/F whether the type is an accepted SQL Server data type.
}
\description{
Check whether a data type is a valid SQL Server data type
}
\details{
The list of accepted SQL types is as follows:
tinyint, smallint, int, bigint, bit, decimal, numeric, money, smallmoney,
float, real, date, time, datetime2, datetimeoffset, datetime, smalldatetime,
char, varchar, text, nchar, nvarchar, ntext.

Does not check numeric components of the type. For example, for input
\code{type = VARCHAR(5)}, only confirms the \code{VARCHAR} component.
}
