% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_support.R
\name{load_control_file}
\alias{load_control_file}
\title{Auto-detecting which file read command is needed}
\usage{
load_control_file(path_and_file_name)
}
\arguments{
\item{path_and_file_name}{location of the file to read into R}
}
\value{
the file contents as a data.frame
}
\description{
Used to support csv, xls, and xlsx formatted control files.
}
