% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_support.R
\name{provide_example}
\alias{provide_example}
\title{Provide an examples of the tools - copying example files from the package
data into a local folder. Includes example control files and worked examples
of the tool in use.}
\usage{
provide_example(example = NA_character_, folder = ".")
}
\arguments{
\item{example}{The example to copy. If does not match an available example,
then prints out a list of the available examples.}

\item{folder}{Folder to copy the example into. Creates the folder if it does
not already exist.}
}
\value{
The path of any newly created file(s).
}
\description{
Provide an examples of the tools - copying example files from the package
data into a local folder. Includes example control files and worked examples
of the tool in use.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# list all the available examples
provide_example()

# load a specific example
provide_example("summary_control_file", "./example")
\dontshow{\}) # examplesIf}
}
