% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTDA.R
\name{ptda}
\alias{ptda}
\title{Pooling Tobit Models for microbiome differential abundance analysis}
\usage{
ptda(
  otu_table,
  metadata,
  covar,
  adjust = NULL,
  prevalence_cutoff = 0.1,
  depth_cutoff = 1000,
  genes_are_rows = TRUE,
  boot = TRUE,
  boot_replicate = 500,
  n_boot_gene = 100,
  alpha = 0.05
)
}
\arguments{
\item{otu_table}{microbiome abundance table generated from 16S rRNA sequencing or shotgun metagenomic sequencing. All entries are integers}

\item{metadata}{sample metadata dataframe}

\item{covar}{the name of the covariate of interest}

\item{adjust}{the names of confounders to adjust for, NULL if no adjustment}

\item{prevalence_cutoff}{features whose prevalence are smaller than the cutoff will be excluded from analysis}

\item{depth_cutoff}{a sample would be discarded if its sequencing depth is smaller than the threshold}

\item{genes_are_rows}{whether the microbiome genes are on the column or the row of the abundance table}

\item{boot}{whether to use bootstrap to estimate scaling factors for the test statistics, default TRUE}

\item{boot_replicate}{number of bootstrap replicates for each gene to estimate scaling factors}

\item{n_boot_gene}{number of genes to apply bootstrap}

\item{alpha}{the cutoff of the adjusted p values}
}
\value{
the reference feature set, the identified DA features and the p values for all the features
}
\description{
Pooling Tobit Models for microbiome differential abundance analysis
}
