% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{readLines_utf8}
\alias{readLines_utf8}
\title{Read lines with requirements of UTF-8 encoding}
\usage{
readLines_utf8(file_name_and_path)
}
\arguments{
\item{file_name_and_path}{A character string giving the path to the file.}
}
\value{
The lines of the file read by \code{readLines} if the file is UTF-8
encoded.#' Otherwise a warning and error about the encoding.
}
\description{
R assumes UTF-8 characters by default. When reading text files there may be
characters that are not part of the UTF-8 encoding. The best option for these
files is to notify the user as early as possible and get them to change the
file.
}
\details{
Continuing to process with characters that are not UTF-8 tends to produce
uninformative error messages.
}
