% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_run.R
\name{run_assembly}
\alias{run_assembly}
\title{Notes
\itemize{
\item documentation detail bullet points to do
}}
\usage{
run_assembly(
  control_file,
  db_connection,
  master_table,
  debug_folder = NA_character_
)
}
\arguments{
\item{control_file}{a data frame containing assembly instructions. Most
likely read into memory by \code{load_control_file}.}

\item{db_connection}{A connection to the database where assembly is to occur.}

\item{master_table}{The name of the table onto which columns should be
assembled. This table must already exist in the database. It is recommended
using the full table name: database.schema.table}

\item{debug_folder}{an existing folder where debug information should be
written to disc. If NA (the default) not debug information is written.}
}
\value{
None, makes permanent changes to \code{master_table} in the database.
}
\description{
Execute assembly tool.
}
\details{
The best way to understand the assembly process is to review a worked example.
Try \code{provide_example} for worked examples, or for example control files.

For each row in the control file, the assembly tool appends a column to
the master table according to the specified instructions. If the column
already exists it is overwritten.

Control files are validated prior to execution. To validate a control file
without execution use \code{validate_assembly_control_file}.

The accepted columns for the control file are:
*

Anywhere dynamic input is accepted in the control file, you can instead give
input in \{curry brackets\}. This input is treated as SQL code.

The intended use of this feature is for making minor adjustments to
variables. For example setting zero values to missing or adjusting dates.
}
