% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_run.R
\name{run_assembly}
\alias{run_assembly}
\title{Execute assembly tool.}
\usage{
run_assembly(
  control_file,
  sheet = NULL,
  db_connection,
  master_table,
  sql_folder = NA_character_,
  debug_folder = NA_character_
)
}
\arguments{
\item{control_file}{location of the control file containing assembly
instructions to read into R. Accepted \code{.csv} and \code{.xlsx} file formats.}

\item{sheet}{Sheet to read if control file is \code{.xlsx} format. As per
\code{openxlsx2::read_xlsx}: either a string (name of a sheet), or an integer
(the position of the sheet). Defaults to the first sheet otherwise.}

\item{db_connection}{A connection to the database where assembly is to occur.}

\item{master_table}{The name of the table onto which columns should be
assembled. This table must already exist in the database. It is recommended
using the full table name: database.schema.table}

\item{sql_folder}{Optional folder location containing SQL scripts. If given
and tables listed in the control file can not be found in the database, then
will check for evidence of table and column names in this folder.}

\item{debug_folder}{an existing folder where debug information should be
written to disc. If NA (the default) not debug information is written.}
}
\value{
None, makes permanent changes to \code{master_table} in the database.
}
\description{
Execute assembly tool.
}
\details{
The best way to understand the assembly process is to review a worked example.
Try \code{provide_example} for worked examples, or for example control files.

For each row in the control file, the assembly tool appends a column to
the master table according to the specified instructions. If the column
already exists it is overwritten.

Control files are validated prior to execution. To validate a control file
without execution use \code{validate_assembly_control_file}.

The accepted columns for the control file are:
\itemize{
\item Enabled - Optional but recommended column, containing True / False or
yes / no. Allows for rows of the control file to be turned on and off.
\item Description  - Optional but recommended column, for user readable
description of the intent of the output column. Not used during assembly.
\item Population_uid  - The name of the identity/uid column in the master table
to use for assembly, delimited with [].
\item Period_start  - The start of the period, can be a column of the master
table, a constant, or SQL code that generates the data.
\item Period_end  - The end of the period, can be a column of the master table,
a constant, or SQL code that generates the data.
\item Measure_file  - The name of the SQL file that prepares the measure. Useful
for tracking the lineage of a measure. If the measure table or columns are
not found in the database, then the file is checked during validation.
\item Measure_table  - The name of the SQL table from which the measure is drawn.
\item Measure_uid  - The name of the identity/uid column in the measure table,
delimited with [].
\item Measure_start  - The start of the measure event, can be a column of the
measure table, a constant, or SQL code that generates the data.
\item Measure_end  - The end of the measure event, can be a column of the measure
table, a constant, or SQL code that generates the data.
\item Measure_value  - The value that should be summarised to create a column in
the master table, can be a column of the measure table, a constant, or SQL
code that generates the data.
\item Output_name  - The name of the output column, should be unique and
delimited with "". For ease of subsequent use it is best to use underscore
instead of spaces in these names.
\item Output_method  - How the measure should be summarised to create the column
for the master table.
\item Output_type  - The SQL data type for the new column.
\item Notes - free text column for adding notes to the control file. This column
is ignored during assembly and does not effect output. Any other column names
are also ignored, but generate a warning.
}

For all inputs, SQL objects (like table and column names) should be delimited
with \verb{[]}; constants should be delimited with \code{""}, and dynamic input should
be delimited with \code{{}}. Dynamic input is treated as SQL code.

The intended use of this feature is for making minor adjustments to
variables. For example setting zero values to missing or adjusting dates.
}
