% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidential_run.R
\name{run_confidential}
\alias{run_confidential}
\title{Notes}
\usage{
run_confidential(control_file, tbl, stable_above = 30)
}
\arguments{
\item{control_file}{a data frame containing confidentialisation instructions.
Most likely read into memory by \code{load_control_file}.}

\item{tbl}{a data frame to confidentialise. Should be in local R memory.}

\item{stable_above}{Minimum source value for which consistent seeds are
required. Defaults to 30, because in practice smaller values come from
different sources.}
}
\value{
A data frame created by applying to \code{tbl} the confidentialisation
rules defined by \code{control_file}.
}
\description{
Execute confidentialisation tool.
}
\details{
The best way to understand the confidentiality process is to review a worked
example. Try \code{provide_example} for worked examples, or for example control
files.

Control files are validated prior to execution. To validate a control file
without execution use \code{validate_confidential_control_file}.

All \code{control_file} values are optional. Where no value is provided, no action
is taken.

The first column of \code{control_file} should contain the command type as listed
below. The remaining columns should match the column names of \code{tbl}. The
accepted commands for the first column are:
\itemize{
\item DROP - where this value is true, the column will be dropped from the
output. Main use case is to remove unneeded label columns.
\item RENAME - text given in this row is used to rename columns. No checks are
applied to ensure the text makes sensible column names.
\item MISSING_TO - numeric value to replace missing values in the column with.
Main use case is to set default values for entity count columns when entity
counts are not applicable.
\item ROUND - the type of rounding to apply. Accepted options are: RR3 and GRR
(for random rounding and graduated random rounding), and CONV10, CONV100, and
CONV1000 (for conventional rounding to base 10, 100, or 1000).
\item SUPPRESS - when to apply suppression, an instruction of the form:
'column-name < value'. Accepts input column names and renamed columns.
Multiple SUPPRESS rows are accepted, where there are multiple suppression
conditions add one per row.
\item NOTES - this row is ignored and does not effect output. It is
intended for adding notes to the control file. You can have as many note rows
as you want.
}

Columns that have had rounding or suppression applied will appear twice in
the output: once with their original values and a second time with rounded /
suppressed values. This second column is named with the prefix "conf_".

Random rounding and Graduated random rounding via \code{run_confidential} include
handling of stable rounding and rounding consistent with thresholds.
\itemize{
\item Stable rounding - means that each unique value in the same column greater
than \code{stable_above} is rounded with the same random seed. This is a more
conservative way of ensuring that counts of identical groups of people are
always rounded the same way.
\item Rounding consistent with thresholds - ensures that where a column is used
for suppression and is also randomly rounded, then values at least/below a
suppression threshold remain at least/below the threshold after rounding.
This prevents the presence/absence of suppression being used to recover the
the raw unrounded value.
}
}
