% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_run.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Execute pipeline tool.}
\usage{
run_pipeline(
  control_file,
  sheet = NULL,
  db_connection_string = NA_character_,
  delay_minutes = 60,
  injection_r = list(),
  injection_sql = list(),
  ignore_warnings = FALSE,
  sink_file = NULL
)
}
\arguments{
\item{control_file}{location of the control file containing pipeline
instructions to read into R. Accepted \code{.csv} and \code{.xlsx} file formats.}

\item{sheet}{Sheet to read if control file is \code{.xlsx} format. As per
\code{openxlsx2::read_xlsx}: either a string (name of a sheet), or an integer
(the position of the sheet). Defaults to the first sheet otherwise.}

\item{db_connection_string}{A connection string for connecting to the
database. Only required if SQL files are included in the pipeline.}

\item{delay_minutes}{Number of minutes to delay execution. Defaults to sixty.
Designed for running pipeline out of hours.}

\item{injection_r}{A list containing named values. For each R script in the
pipeline, a separate environment will be created and populated with these
variables before the file is executed (see details).}

\item{injection_sql}{A list containing named values. For each SQL script in
the pipeline, where the names are found in the SQL code, these will be
replaced with their values (see details).}

\item{ignore_warnings}{T/F whether execution of R or SQL scripts should
continue even if warnings occur. If \code{FALSE} (the default) then if a warning
occurs during the execution of any file in the pipeline, execution of that
file will stop. If \code{TRUE} then will suppress all warnings, files in the
pipeline will only stop running if they encounter an error.}

\item{sink_file}{If provided, a file to sink console progress reports to.}
}
\value{
A data frame containing all the files run in order, along with start
and end times and the status of their execution (completed, stopped with
warning or error). If a warning or error occurred, then status also includes
the warning or error message.
}
\description{
Execute pipeline tool.
}
\details{
The best way to understand the pipeline tool is to review a worked example.
Try \code{provide_example} for worked examples, or for example control files.

For each row in the control file, the pipeline tool executes the file using
either \code{try_run_R_file} or \code{try_run_SQL_file}, depending on the file
extension.

Control files are validated prior to execution. To validate a control file
without execution use \code{validate_pipeline_control_file}.

The accepted columns for the control file are:
\itemize{
\item Enabled - Optional but recommended column, containing True / False or
yes / no. Allows for rows of the control file to be turned on and off.
\item Order - Optional column for setting the order of the files to be run.
Files without an order are run in the order they appear in the control file
after all files with an order.
\item Folder - The folder that contains the file to be executed. All folders are
processed with \code{adjust_file_path_handling} which converts format to match
R and data lab expectations.
\item File - The file to be executed. Only R and SQL scripts are accepted.
\item Notes - free text column for adding notes to the control file. This column
is ignored during execution and does not effect output. Any other column
names are also ignored, but generate a warning.
}

As an alternative, the text \verb{STOP IF ANY FAILURES} can be entered can be
entered in the \code{File} column of the control file. This is a special command
and will cause the tool to stop if any failures have occurred. This is
useful for separating stages of the pipeline where subsequent stages should
run only if all of the previous stage(s) complete.

If the \code{delay_minutes} argument is at least 60, then a random additional
delay of 1-60 minutes is added upon execution. This is to prevent all the
pipelines executed at 5pm from causing strain on the system when they all
attempt to run one hour later at 6pm.

\code{injection_r} provides a way to insert dynamic values into the environment
when R scripts run. For example \code{injection_r = list(ext = "csv")} would be
equivalent to adding the code \code{ext <- "csv"} at the top of every R script in
the pipeline. Case sensitive. \code{injection_r} exists to allow parameters to be
passed to scripts.

\code{injection_sql} provides a way to insert dynamic values into an SQL script.
It is the equivalent of SQL CMD mode for the pipeline tool (as SQL CMD mode
does not work as expected via ODBC connection). Case sensitive. For example
\code{injection_sql = list("$(tbl)" = "[my_table]")} would replace all instances
of \verb{$(tbl)} with \verb{[my_table]} in all SQL scripts. \code{injection_sql} exists to
allow parameters to be passed to scripts.
}
