% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_run.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Execute pipeline tool.}
\usage{
run_pipeline(
  control_file,
  sheet = NULL,
  db_connection_string = NA_character_,
  delay_minutes = 0,
  ignore_warnings = FALSE
)
}
\arguments{
\item{control_file}{location of the control file containing pipeline
instructions to read into R. Accepted \code{.csv} and \code{.xlsx} file formats.}

\item{sheet}{Sheet to read if control file is \code{.xlsx} format. As per
\code{openxlsx2::read_xlsx}: either a string (name of a sheet), or an integer
(the position of the sheet). Defaults to the first sheet otherwise.}

\item{db_connection_string}{A connection string for connecting to the
database. Only required if SQL files are included in the pipeline.}

\item{delay_minutes}{Number of minutes to delay execution. Defaults to zero.
Useful if wanting to run pipeline out of hours.}

\item{ignore_warnings}{T/F whether execution of R or SQL scripts should
continue even if warnings occur. If \code{FALSE} (the default) then if a warning
occurs during the execution of any file in the pipeline, execution of that
file will stop. If \code{TRUE} then will suppress all warnings, files in the
pipeline will only stop running if they encounter an error.}
}
\value{
A data frame containing all the files run in order, along with start
and end times and the status of their execution (completed, stopped with
warning or error). If a warning or error occured, then status also includes
the warning or error message.
}
\description{
Execute pipeline tool.
}
\details{
The best way to understand the pipeline tool is to review a worked example.
Try \code{provide_example} for worked examples, or for example control files.

For each row in the control file, the pipeline tool executes the file using
either \code{try_run_R_file} or \code{try_run_SQL_file}, depending on the file
extension.

Control files are validated prior to execution. To validate a control file
without execution use \code{validate_pipeline_control_file}.

The accepted columns for the control file are:
\itemize{
\item Enabled - Optional but recommended column, containing True / False or
yes / no. Allows for rows of the control file to be turned on and off.
\item Order - Optional column for setting the order of the files to be run.
Files without an order are run in the order they appear in the control file
after all files with an order.
\item Folder - The folder that contains the file to be executed. All folders are
processed with \code{adjust_file_path_handling} which converts format to match
R and data lab expectations.
\item File - The file to be executed. Only R and SQL scripts are accepted.
\item Notes - free text column for adding notes to the control file. This column
is ignored during execution and does not effect output. Any other column
names are also ignored, but generate a warning.
}

As an alternative, the text \verb{STOP IF ANY FAILURES} can be entered can be
entered in the \code{File} column of the control file. This is a special command
and will cause the tool to stop if any failures have occurred. This is
useful for separating stages of the pipeline where subsequent stages should
run only if all of the previous stage(s) complete.
}
