% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_run.R
\name{run_summary}
\alias{run_summary}
\title{Notes
\itemize{
\item awaiting helper documentation
\item awaiting tests
}}
\usage{
run_summary(control_file, tbl, remote_na_from_groups = TRUE)
}
\description{
Notes
\itemize{
\item awaiting helper documentation
\item awaiting tests
}
}
