% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_run.R
\name{run_summary}
\alias{run_summary}
\title{Notes
\itemize{
\item awaiting helper documentation
\item awaiting tests
\item remote_na_from_groups parameter needs to be integrated
\item debug reporting needed
\item deliberate casting of outputs to required types (character or numeric)
}}
\usage{
run_summary(
  control_file,
  tbl,
  remote_na_from_groups = TRUE,
  debug_folder = NA_character_
)
}
\arguments{
\item{control_file}{a data frame containing summary instructions. Most likely
read into memory by \code{load_control_file}.}

\item{tbl}{a data frame to summarise. Can be in-memory or remote accessed
with dbplyr.}

\item{remote_na_from_groups}{T/F whether missing values from grouping columns
should be excluded from the results. Defaults to TRUE.}

\item{debug_folder}{an existing folder where debug information should be
written to disc. If NA (the default) not debug information is written.}
}
\value{
A data frame from executing the summary defined by each row of the
control file and appending all the summaries together.
}
\description{
Execute summary tool.
}
\details{
The following checks are run and generate a failure if not passed:
\itemize{
\item required columns exist in data frame
\item dynamic formula can be executed
\item Each row has at least one summary generated
}

The following checks are run and only generate a warning if not passed:
\itemize{
\item acceptable column names ("enabled", "group", "label", "distinct", "count",
"sum", "entity", "stddev", "notes")
\item column is not empty
\item columns to sum are numeric
}
}
