% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_run.R
\name{run_summary}
\alias{run_summary}
\title{Notes}
\usage{
run_summary(
  control_file,
  tbl,
  remove_na_from_groups = TRUE,
  debug_folder = NA_character_
)
}
\arguments{
\item{control_file}{a data frame containing summary instructions. Most likely
read into memory by \code{load_control_file}.}

\item{tbl}{a data frame to summarise. Can be in-memory or remote accessed
with dbplyr.}

\item{remove_na_from_groups}{T/F whether missing values from grouping columns
should be excluded from the results. Defaults to TRUE.}

\item{debug_folder}{an existing folder where debug information should be
written to disc. If NA (the default) not debug information is written.}
}
\value{
A data frame from executing the summary defined by each row of the
control file and appending all the summaries together.
}
\description{
Execute summary tool.
}
\details{
For each row in the control file, the summary tool produces a summary of tbl
according to the specified instructions. All these summaries are combined
into a single table and returned.

Validates its control file prior to execution. To validate a control file
without execution use \code{validate_summary_control_file}.

The best way to understand the summary process is to review a worked example.
Try \code{provide_example} for worked examples, or for example control files.
}
