% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_run.R
\name{run_summary}
\alias{run_summary}
\title{Execute summary tool.}
\usage{
run_summary(
  control_file,
  sheet = NULL,
  tbl,
  remove_na_from_groups = TRUE,
  save_file = NA_character_,
  debug_folder = NA_character_
)
}
\arguments{
\item{control_file}{location of the control file containing summary
instructions to read into R. Accepted \code{.csv} and \code{.xlsx} file formats.}

\item{sheet}{Sheet to read if control file is \code{.xlsx} format. As per
\code{openxlsx2::read_xlsx}: either a string (name of a sheet), or an integer
(the position of the sheet). Defaults to the first sheet otherwise.}

\item{tbl}{a data frame to summarise. Can be in-memory or remote accessed
with dbplyr.}

\item{remove_na_from_groups}{T/F whether missing values from grouping columns
should be excluded from the results. Defaults to TRUE.}

\item{save_file}{specify the file to save results to. Optional, overides the
FILE column of the control file if provided.}

\item{debug_folder}{an existing folder where debug information should be
written to disc. If NA (the default) no debug information is written.}
}
\value{
A data frame from executing the summary defined by each row of the
control file and appending all the summaries together.
}
\description{
Execute summary tool.
}
\details{
The best way to understand the summary process is to review a worked example.
Try \code{provide_example} for worked examples, or for example control files.

For each row in the control file, the summary tool produces a summary of tbl
according to the specified instructions. All these summaries are combined
into a single table and returned.

Control files are validated prior to execution. To validate a control file
without execution use \code{validate_summary_control_file}.

The accepted columns for the control file are:
\itemize{
\item ENABLED - TRUE/FALSE - If this column is included then rows set to FALSE
will be omitted. Except for ENABLED, you can have any number of columns of
each type.
\item FILE - file path and name - where the summary output should be saved,
output can be divided across multiple files. Files and folders will be
created if they do not exist, and overwritten if they do exist.
\item LABEL - free text - allows you to include arbitrary text in your output,
this is intended to provide a description of the summary.
\item GROUP - name of columns of \code{tbl} - When the summary for the row is
produced, the output is grouped by the group columns.
\item DISTINCT - name of columns of \code{tbl} - counts the number of distinct values
in the column.
\item COUNT - name of columns of \code{tbl} - counts the number of non-missing values
in the column.
\item SUM - name of columns of \code{tbl} - calculates the sum total of the column,
during validation the tool will error is this column is not numeric.
\item STDDEV - name of columns of \code{tbl} - calculates the standard deviation of the
column.
\item ENTITY - name of columns of \code{tbl} -  similar to DISTINCT, but checks for
columns with \verb{*__min} and \verb{*__max} suffixes and takes a distinct over the union
of these columns if available. Designed for counting entities.
\item SEED - name of columns of \code{tbl} - similar to SUM, but takes remainder by
100 first. Intended to be applied to ID columns to generate a value that
can serve as a seed for consistent rounding.
\item WHERE - a dynamic formula in \{curly brackets\} to filter the input table
before summarising.
\item NOTES - free text - column is ignored and does not effect output. Intended
for adding notes to control file. Any other column names are also ignored,
but generate a warning.
}

Anywhere you can give the name of a column of \code{tbl} in the control file,
you can instead give input in \{curly brackets\}. This input is treated as R
code. If \code{tbl} is a remote data frame, then this will be translated to SQL
using dbplyr.

The intended use of this feature is for making minor adjustments to
variables. For example replacing missing values.
}
