% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{run_time_inform_user}
\alias{run_time_inform_user}
\title{Prints to console time of function call followed by a message.}
\usage{
run_time_inform_user(msg, log = NA_character_)
}
\arguments{
\item{msg}{A message to display time stamped for the user.}

\item{log}{Optional path to an existing log to write to.}
}
\value{
The time-stamped message invisibly.
}
\description{
Prints to console time of function call followed by a message.
}
