% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universal_support.R
\name{save_code_to_script}
\alias{save_code_to_script}
\title{Save code to file for debugging}
\usage{
save_code_to_script(query, desc, folder_path)
}
\arguments{
\item{query}{the text of the query to save. This may be generated using
\code{dplyr::show_query} where required.}

\item{desc}{a description of the query. Use to name the file. If an
extension is included this extension is used. Otherwise .txt is used.}

\item{folder_path}{The path to save the query. If provided will attempt to
save a copy of the query to the provided folder. Errors if directory does
not exist.}
}
\value{
The location and name of the saved file.
}
\description{
For transparency and ease of debugging, this function simplifies writing to
disc key code components. It provides a standardized way to save SQL code
or character strings using within dynamic dplyr code.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
code = "SELECT * FROM mytable"
written_file = save_code_to_script(code, "saved code", ".")
\dontshow{\}) # examplesIf}
}
