% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_support.R
\name{sql_cmd_mode_fix}
\alias{sql_cmd_mode_fix}
\title{Replace SQL command mode variables with their values}
\usage{
sql_cmd_mode_fix(code_string)
}
\arguments{
\item{code_string}{A character vector comprising sql code}
}
\value{
A character vector comprising sql code with any command mode ":SETVAR"
statements removed, and variables replaced with values
}
\description{
This function searches SQL code to identify instances where SQL CMD Mode
variables have been set and replaces these with their values. The intention
is to enable the code to be called from R which does not enable CMD Mode.
Any :SETVAR statements are replaced with empty space, but new lines are
preserved.
}
