% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidential_support.R
\name{suppression_format_extract}
\alias{suppression_format_extract}
\title{Decompose suppression instructions into individual components and validate
correctness of instructions.}
\usage{
suppression_format_extract(string)
}
\arguments{
\item{string}{a text string with the suppression instructions to process.}
}
\value{
a list with five components:
\itemize{
\item \code{input} the original input string.
\item \code{column} the column to process.
\item \code{sign} the inequality that defines suppression.
\item \code{threshold} the numeric value to suppress below.
\item \code{valid} T/F whether the decomposed input makes for valid suppression rules.
}
}
\description{
Decompose suppression instructions into individual components and validate
correctness of instructions.
}
\details{
Splits an input string into three components and then evaluates whether these
components make valid suppression instructions.

For example: "colname < 6" will be decomposed into column = "colname", sign
= "<", and threshold = "6".
}
