% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_support.R
\name{tolower_control_file_cells}
\alias{tolower_control_file_cells}
\title{Column names in control file to lower case}
\usage{
tolower_control_file_cells(summary_control_file, tbl_cols)
}
\arguments{
\item{summary_control_file}{a data frame containing summary instructions.}

\item{tbl_cols}{an array with the column names of the table. Where these
are found in the control file set them to lower case.}
}
\value{
The control file with cells contents set to lower case to match
lower case names of \code{tbl_cols}. Affects columns of type group, distinct,
count, sum, entity, stddev, and where.

This internal function exists because R is case sentitive, but SQL is not,
and control files might not be case sentivie either.
}
\description{
Column names in control file to lower case
}
