% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_validate.R
\name{validate_assembly_control_file}
\alias{validate_assembly_control_file}
\title{Confirm that control file instructions can be executed.}
\usage{
validate_assembly_control_file(
  control_file,
  db_connection,
  master_table,
  sql_folder = NA_character_
)
}
\arguments{
\item{control_file}{a data frame containing assembly instructions. Most
likely read into memory by \code{load_control_file}.}

\item{db_connection}{A connection to the database where assembly is to occur.}

\item{master_table}{The name of the table onto which columns should be
assembled. It is recommended using the full table name: database.schema.table}

\item{sql_folder}{Optional folder location containing SQL scripts. If given
and tables listed in the control file can not be found in the database, then
will check for evidence of table and column names in this folder.}
}
\value{
T/F whether or not all validation checks are passed. Generating
warnings for all failed checks.
}
\description{
Confirm that control file instructions can be executed.
}
\details{
The following checks are run and generate a failure if not passed:
\itemize{
\item Required column names are present.
\item Inputs match expected formats / delimiters.
\item Dynamic inputs contain \code{no_obvious_escaping_injection}
\item Master table exists.
\item Master table columns exist.
\item Measure tables exist in the database or are found in the measure_file.
\item Measure columns exist in the database or are found in the measure_file.
\item Output methods are accepted types.
\item Output types are accepted SQL data types
}

The following checks are run and only generate a warning if not passed:
\itemize{
\item Unrecognised column names are present.
}
}
