% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidential_validate.R
\name{validate_confidential_control_file}
\alias{validate_confidential_control_file}
\title{Confirm that control file instructions can be executed.}
\usage{
validate_confidential_control_file(control_file, tbl)
}
\arguments{
\item{control_file}{a data frame containing confidentialisation instructions.
Most likely read into memory by \code{load_control_file}.}

\item{tbl}{a data frame to confidentialise. Should be in local R memory.}
}
\value{
T/F whether or not all validation checks are passed. Generating
warnings for all failed checks.
}
\description{
Confirm that control file instructions can be executed.
}
\details{
The following checks are run and generate a failure if not passed:
\itemize{
\item First column of \code{control_file} contains confidentiality commands
\item Remaining columns listed in \code{control_file} match the column names of \code{tbl}.
\item At most one row of DROP, RENAME, ROUND, and MISSING_TO instructions.
\item Rounding instructions are of the accepted types (RR3, GRR, CONV10, CONV100,
CONV1000).
\item Missing NA value options are numeric.
\item Suppression rules match the accepted pattern (column_name < number value).
}
}
