% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_validate.R
\name{validate_pipeline_control_file}
\alias{validate_pipeline_control_file}
\title{Confirm that pipeline control file instructions can be executed.}
\usage{
validate_pipeline_control_file(control_file, db_connection)
}
\description{
Confirm that pipeline control file instructions can be executed.
}
