% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_validate.R
\name{validate_summary_control_file}
\alias{validate_summary_control_file}
\title{Notes}
\usage{
validate_summary_control_file(control_file, tbl)
}
\arguments{
\item{control_file}{a data frame containing summary instructions. Most likely
read into memory by \code{load_control_file}.}

\item{tbl}{a data frame to summarise. Can be in-memory or remote accessed
with dbplyr.}
}
\value{
T/F whether or not all validation checks are passed. Generating
warnings for all failed checks.
}
\description{
Confirm that control file instructions can be executed.
}
\details{
The following checks are run and generate a failure if not passed:
\itemize{
\item required columns exist in data frame
\item dynamic formula can be executed
\item Each row has at least one summary generated
\item columns to sum are numeric
\item grouping columns are not dynamic
}

The following checks are run and only generate a warning if not passed:
\itemize{
\item acceptable column names ("enabled", "group", "label", "distinct", "count",
"sum", "entity", "stddev", "notes")
\item column is not empty
\item grouping columns not used for summarising
}
}
